Refactor and enhance session management and UI components
- Removed unused labelOf function from overlap-audit.spec.ts.
- Improved session handling in smoke.spec.ts to verify session ID changes on restart and ensure proper session closure.
- Added session reset functionality in software-only-recipes.spec.ts to maintain a clean state between tests.
- Updated UI issues audit to remove unnecessary brake transmission checks.
- Enhanced Bench component in App.tsx to manage synthetic peers and their states, including starting and stopping stimuli.
- Introduced verification runner in Diagnostics component to execute and display test results.
- Added session management functions in Settings component to handle session restarts and lease releases.
- Updated DriveConsole to use callbacks for disarming controls and managing keyboard focus.
- Improved API interactions in api.ts to support synthetic peers and tests.
- Updated documentation for firmware flashing and GPIO audit to reflect new board configurations and requirements.
- Fixed CMake and test files to ensure compatibility with the new board definitions and configurations.
- Created new board definitions for esp32-s3-devkitc-1-n16r8 to support enhanced hardware capabilities.

+++ native-test/test/test_signal_chains.cpp
@@ -66,7 +66,7 @@ sts.light_left=sts.light_right=sts.light_brake=sts.light_head=true;
 protocol::Frame f;generated::encode(sts,f);CHECK_EQ(f.dlc,3,"DLC=3");
 generated::SysSafetySts ds;generated::decode(f.view(),ds);
 CHECK_EQ(ds.estop_active,true,"estop_active=true");CHECK_EQ(ds.heartbeat_ok,true,"heartbeat_ok=true");
-uint8_t light_state=ds.light_left|(ds.light_right<<1)|(ds.light_brake<<2)|(ds.light_head<<3);
+uint8_t light_state=static_cast<uint8_t>(ds.light_left)|(static_cast<uint8_t>(ds.light_right)<<1)|(static_cast<uint8_t>(ds.light_brake)<<2)|(static_cast<uint8_t>(ds.light_head)<<3);
 CHECK_EQ(light_state,0x0F,"light_state=0x0F");}
 
 int main(){printf("=== Stage 2: CAN Signal Chain Tests ===\n");
